Fix variation in line size.
When drawing the board, some lines would randomly be thicker
(varying between one and two pixels width).

CCMAIL: sasch.pe@gmx.de


svn path=/trunk/KDE/kdegames/kigo/; revision=1268331

+++ src/gui/graphicsview/gamescene.h
@@ -83,7 +83,7 @@ private:
     QRectF m_boardRect;                     ///< Position of board in the scene
     QRectF m_mouseRect;                     ///< Board mouse interaction rect
     QRectF m_gridRect;                      ///< Board grid rect
-    qreal m_cellSize;                       ///< Width of board grid cell
+    int m_cellSize;                         ///< Width of board grid cell
     QSize m_stonePixmapSize;                ///< Size of Go stone pixmap
     QSize m_placementMarkerPixmapSize;
     int m_boardSize;                        ///< Go board size (9, 13, 19, ..)
